add asteroids feature
