+++ Cargo.toml
@@ -15,3 +15,4 @@ serde_json = "1.0.117"
 [features]
 radius = []
 azerty = []
+asteroids = []
